Add testing infrastructure and refactor
This commit adds the infrastructure we will use for testing, and writes
a few simple tests of basic routes to check its working properly.

Additionally, we move all of the routes over to `views.py` using
`flask.Blueprint`, in order to split the code into smaller components.
The written tests reflect this change.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>