Sketch outline for the webhook
This commit sketches the outline for the webhook which will receive the
contents of text messages from Twilio, write the texter to the database,
figure out the proper response, and then send it.

Not all of these methods are filled in nor are they all tested, but this
is committed so we begin divying up the pieces.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>