Update design using html5 template
Now we have an actual fairly nice looking design for the website. This
is just a static page (everything dynamic happens through the text
message).

When making changes to `main.css` in the future, make sure to update the
`?v=` take in `layout.html` to force a cache update. We set `v` to be
`git rev-parse HEAD`.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>